TAP5-427: Allow injection of Tapestry services into Spring beans
git-svn-id: https://svn.eu.apache.org/repos/asf/tapestry/tapestry5/trunk@730009 13f79535-47bb-0310-9956-ffa450edef68

+++ tapestry-core/src/main/java/org/apache/tapestry5/internal/services/RenderQueueImpl.java
@@ -89,7 +89,7 @@ public class RenderQueueImpl implements RenderQueue
         long endNanos = System.nanoTime();
 
         long elapsedNanos = endNanos - startNanos;
-        double elapsedSeconds = ((float) elapsedNanos) / 1000000000F;
+        double elapsedSeconds = ((double) elapsedNanos) / 1000000000d;
 
         logger.debug(TapestryMarkers.RENDER_COMMANDS,
                      String.format("Executed %,d rendering commands (max queue depth: %,d) in %.3f seconds",
